Refactor test imports to align with new Pro directory structure
- Updated import paths in multiple test files to reflect the relocation of components to the Pro directory.
- This change ensures consistency and maintains the integrity of the testing framework following recent structural adjustments.
