+++ node_package/tests/serverRenderReactComponent.test.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import serverRenderReactComponent from '../src/serverRenderReactComponent.ts';
-import * as ComponentRegistry from '../src/ComponentRegistry.ts';
+import * as ComponentRegistry from '../src/pro/ComponentRegistry.ts';
 import type {
   RenderParams,
   RenderResult,
